Updated to include wrench size
If you have your Trinus inside an enclosure, an adjustable wrench won't fit.  Meanwhile, folks in the U.S. (at least) won't have a complete set of small metric wrenches -- so I added the size of the locking nuts, so nobody has to buy a bunch of wrenches to get one in the correct size.